feat: added the UseInvoker module
The UseInvoker module can be used for generating new modules
which itself gets a using directive that accepts an atom as first
argument.

If the `required_utils` option is passed then the calling module also
invokes the passed function names (that must be arguments) to ensure
that all the specified core utilities will be available.