refactor(doctester): accept options instead of a list of modules
BREAKING CHANGE:

Refactored the `HypeLib.DocTester` module from accepting a list of
modules to a keyword list when the `__using__` macro is invoked.

This allows us more control about how the DocTester should behave.

Therefore we have implemented the `modules` and `async` options.

The `modules` option is used for passing the modules to test to the
DocTester.

The `async` option is used for controlling the ExUnit prelude which is
generated for the calling module. It gets passed down to the `use
ExUnit.Case` call and tells ExUnit if it should run all the tests
parallel or not.